+++ src/components/Onfido/index.desktop.js
@@ -1,9 +1,9 @@
-import BaseOnfido from './BaseOnfido';
+import BaseOnfidoWeb from './BaseOnfidoWeb';
 import onfidoPropTypes from './onfidoPropTypes';
 
 // On desktop, we do not want to teardown onfido, because it causes a crash.
 // See https://github.com/Expensify/App/issues/6082
-const Onfido = BaseOnfido;
+const Onfido = BaseOnfidoWeb;
 
 Onfido.propTypes = onfidoPropTypes;
 Onfido.displayName = 'Onfido';
